backend/pyproject.toml
[project]
name = "home-site-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "python-dotenv>=1.1",
    "alembic>=1.15",
    "psycopg2-binary>=2.9",
]
